Add Triggers
The class is extracted from the Network class. It's a Seperation of Concerns
thing. In the end this will make it easier to change the network module.
Note that the lock used here is now specific to the callbacks, it used to be
shared with the lock used in sending API calls. I think this is clearer.
